push conflict can happen

+++ src/app/main/main.component.ts
@@ -69,7 +69,7 @@ export class MainComponent implements OnInit {
       });
 
       this.loginService.getAllByReceiver(this.cookieUsername).subscribe(res => {
-        res.array.forEach(element => {
+        res.forEach(element => {
           this.handleMessage(element);
         });
       });
@@ -164,7 +164,8 @@ export class MainComponent implements OnInit {
   toggleNotificationPanel() {
     this.openNotificationPanel = !this.openNotificationPanel;
     this.notifications_count = 0;
-    this.loginService.deleteByReceiver(this.cookieUsername).subscribe(res => {
+    if(this.openNotificationPanel == false)
+      this.loginService.deleteByReceiver(this.cookieUsername).subscribe(res => {
 
     });
   }
